REFACTO(HomeScreen.kt) : 이번주 학습 성취율 스크린 이동
