+++ presentation/src/main/java/com/capston/presentation/ui/HomeScreen.kt
@@ -6,8 +6,10 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Build
 import androidx.annotation.RequiresApi
+import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.LinearEasing
+import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Canvas
@@ -117,8 +119,14 @@ import androidx.compose.material3.DatePickerDialog
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.TextButton
 import androidx.compose.material3.rememberDatePickerState
+import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.withStyle
+import com.capston.domain.response.enum_class.DayOfWeek
 import com.capston.domain.response.plan.LectureAliasResponse
+import com.capston.presentation.theme.LightGray5
 import com.capston.presentation.theme.WarmPurple
 import java.text.SimpleDateFormat
 import java.util.Date
@@ -160,6 +168,9 @@ fun HomeScreen(homeViewModel: HomeViewModel, planViewModel: PlanViewModel) {
     var examTitle by rememberSaveable { mutableStateOf("예정된 계획이 없어요.") }
     var examDate by rememberSaveable { mutableStateOf("2025-05-16") } // 예시 날짜
     var dDay by rememberSaveable { mutableStateOf(0) } // D-Day 계산된 값
+
+    // 이번주 학습 성취율 섹션 확장/축소 상태
+    var isWeeklyExpanded by remember { mutableStateOf(true) }
 
     Scaffold(modifier = Modifier.fillMaxSize()) {
         Column(
@@ -208,6 +219,25 @@ fun HomeScreen(homeViewModel: HomeViewModel, planViewModel: PlanViewModel) {
                     modifier = Modifier.weight(1f)
                 )
             }
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            // 이번주 학습 성취율 섹션 추가
+            Card(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp, vertical = 8.dp),
+                shape = RoundedCornerShape(10.dp),
+                colors = CardDefaults.cardColors(containerColor = Color.White),
+                border = BorderStroke(1.dp, color = LightGray60)
+            ) {
+                WeeklyAchievementGraph(
+                    isExpanded = isWeeklyExpanded,
+                    onToggle = { isWeeklyExpanded = it }
+                )
+            }
+
+            Spacer(modifier = Modifier.height(16.dp))
         }
     }
 
@@ -237,6 +267,243 @@ fun HomeScreen(homeViewModel: HomeViewModel, planViewModel: PlanViewModel) {
         }
     )
 }
+
+// ProfileScreen에서 가져온 이번주 학습 성취율 관련 컴포넌트들
+@SuppressLint("DefaultLocale")
+@Composable
+fun WeeklyAchievementGraph(
+    isExpanded: Boolean,
+    onToggle: (Boolean) -> Unit
+) {
+    // 요일별 학습 성취 데이터 (true: 달성, false: 미달성)
+    val weekDaysData = listOf(
+        DayAchievement(DayOfWeek.MON, true),
+        DayAchievement(DayOfWeek.TUE, true),
+        DayAchievement(DayOfWeek.WED, true),
+        DayAchievement(DayOfWeek.THU, false),
+        DayAchievement(DayOfWeek.FRI, true),
+        DayAchievement(DayOfWeek.SAT, false),
+        DayAchievement(DayOfWeek.SUN, true)
+    )
+
+    // 성취율 계산
+    val achievedDays = weekDaysData.count { it.isAchieved }
+    val totalDays = weekDaysData.size
+    val achievementRate = (achievedDays.toFloat() / totalDays) * 100
+
+    // 연속 달성일 계산
+    var currentStreak = 0
+    var maxStreak = 0
+
+    weekDaysData.forEach { dayData ->
+        if (dayData.isAchieved) {
+            currentStreak++
+            maxStreak = maxOf(maxStreak, currentStreak)
+        } else {
+            currentStreak = 0
+        }
+    }
+
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 20.dp, vertical = 16.dp),
+        horizontalAlignment = Alignment.Start
+    ) {
+        // 제목 행 (아이콘 추가)
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(bottom = if (isExpanded) 20.dp else 0.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            // 아이콘
+            Image(
+                painter = painterResource(id = R.drawable.screen_profile_week_completion),
+                contentDescription = "이번주 학습 성취율 아이콘",
+                modifier = Modifier
+                    .size(24.dp)
+                    .padding(end = 5.dp)
+            )
+
+            // 제목
+            Text(
+                text = "이번주 학습 성취율",
+                style = MaterialTheme.typography.titleMedium,
+                color = Color.Black,
+                modifier = Modifier.padding(end = 5.dp)
+            )
+
+            // 오른쪽 화살표 - 클릭 가능한 토글 버튼으로 변경
+            IconButton(
+                onClick = { onToggle(!isExpanded) },
+                modifier = Modifier.size(24.dp)
+            ) {
+                // 회전 애니메이션 추가
+                val rotationState by animateFloatAsState(
+                    targetValue = if (isExpanded) 90f else 0f,
+                    label = "rotationAnimation"
+                )
+
+                Image(
+                    painter = painterResource(id = R.drawable.screen_profile_see_more_iv),
+                    contentDescription = if (isExpanded) "접기" else "펼치기",
+                    modifier = Modifier.rotate(rotationState)
+                )
+            }
+
+            // 나머지 공간 채우기
+            Spacer(modifier = Modifier.weight(1f))
+
+            // 성취율 표시 (펼치기 상태에 관계없이 표시)
+            Text(
+                text = "성취율 ${String.format("%.0f", achievementRate)}%",
+                style = MaterialTheme.typography.bodyMedium,
+                color = Color.Gray,
+                fontWeight = FontWeight.Bold
+            )
+        }
+
+        // 회색 테두리 박스 - 중앙 정렬로 변경
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 10.dp),
+            contentAlignment = Alignment.Center
+        ) {
+            Card(
+                modifier = Modifier
+                    .padding(10.dp),
+                shape = RoundedCornerShape(5.dp), // 둥근 모서리
+                colors = CardDefaults.cardColors(
+                    containerColor = Color.White // 흰색 배경
+                ),
+                border = BorderStroke(1.dp, color = LightGray60)
+            ) {
+                // 연속 달성일 정보 텍스트
+                Text(
+                    buildAnnotatedString {
+                        append("이번주 ")
+                        withStyle(style = SpanStyle(color = MainPurple)) {
+                            append("${achievedDays}일")
+                        }
+                        append(" 학습 목표를 달성했어요")
+                    },
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = Color.Black,
+                    fontSize = 14.sp,
+                    modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp)
+                )
+            }
+        }
+
+        // 확장된 경우에만 요일별 체크 아이콘 표시
+        AnimatedVisibility(visible = isExpanded) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 20.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                // 요일별 체크 표시
+                DayAchievementChecks(weekDaysData = weekDaysData)
+
+                Spacer(modifier = Modifier.height(20.dp))
+
+                // 추가 정보 표시
+                Text(
+                    text = "최대 연속 달성: ${maxStreak}일",
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MainPurple,
+                    fontWeight = FontWeight.Bold
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun DayAchievementChecks(weekDaysData: List<DayAchievement>) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        // 요일별 체크 아이콘 표시를 위해 너비를 제한하여 모든 요일이 표시되도록 함
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 16.dp)
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceEvenly,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                // 각 요일별 아이템
+                weekDaysData.forEach { dayData ->
+                    // 각 아이템은 고정된 너비를 가지도록 설정
+                    Box(
+                        modifier = Modifier.weight(1f),
+                        contentAlignment = Alignment.Center
+                    ) {
+                        DayAchievementItem(
+                            day = dayData.day,
+                            isAchieved = dayData.isAchieved
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun DayAchievementItem(day: DayOfWeek, isAchieved: Boolean) {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+        modifier = Modifier.padding(horizontal = 2.dp) // 여백 줄임
+    ) {
+        // 요일 표시
+        Text(
+            text = day.label,
+            style = MaterialTheme.typography.bodyMedium,
+            fontSize = 14.sp,
+            color = Color.Gray
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        // 체크 아이콘 배경
+        Box(
+            modifier = Modifier
+                .size(36.dp) // 크기 약간 줄임
+                .background(
+                    color = if (isAchieved) MainPurple else LightGray5,
+                    shape = androidx.compose.foundation.shape.CircleShape
+                ),
+            contentAlignment = Alignment.Center
+        ) {
+            // 달성했을 경우에만 체크 아이콘 표시
+            if (isAchieved) {
+                Image(
+                    painter = painterResource(R.drawable.screen_profile_check_iv),
+                    contentDescription = "달성 완료",
+                    modifier = Modifier.size(18.dp) // 아이콘 크기 조정
+                )
+            }
+        }
+    }
+}
+
+// 일별 성취 데이터 클래스
+data class DayAchievement(
+    val day: DayOfWeek,
+    val isAchieved: Boolean
+)
 
 @Composable
 fun LearningStatusCard(
